requirements.txt
streamlit
torch
torchaudio
transformers
diffusers
soundfile
Pillow
huggingface_hub
